+++ python/src/causal_rl_bench/utils/env_utils.py
@@ -22,5 +22,5 @@ def combine_spaces(space_1, space_2):
     upper_bound = np.concatenate((space_1.high, space_2.high))
     return spaces.Box(low=lower_bound,
                       high=upper_bound,
-                      dtype=np.float64)
+                      dtype=np.float32)
